input: rename mp_event_drop_files to mp_input_drop_files
Move it to input.c. Will be used in next commit.
